+++ docs/sdks/web/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 21.2.1
+
+* Add transaction support for Databases and TablesDB
+
 ## 21.1.0
 
 * Deprecate `createVerification` method in `Account` service